lub_backend/requirements.txt
aiofiles==24.1.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrdict==2.0.1
azure-core==1.36.0
azure-storage-blob==12.27.1
babel==2.17.0
bce-python-sdk==0.9.46
bcrypt==4.1.2
beautifulsoup4==4.13.5
black==24.10.0
blinker==1.9.0
boto3==1.40.31
botocore==1.40.31
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.2.1
coverage==7.10.7
cryptography==45.0.7
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fire==0.7.1
flake8==7.3.0
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.59.2
future==1.0.0
greenlet==3.2.4
grep==0.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iniconfig==2.1.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle==3.0.0
pandas==2.3.2
pathspec==0.12.1
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
premailer==3.10.0
protobuf==6.32.1
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.20.2
pyparsing==3.2.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
rarfile==4.2
reportlab==4.4.3
requests==2.32.3
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
scikit-image==0.25.2
scipy==1.13.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
termcolor==3.1.0
tifffile==2025.5.10
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
visualdl==2.5.3
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
